requirements.txt
telethon
python-dotenv
emoji
requests
beautifulsoup4
groq
streamlit
pandas
streamlit-option-menu
